同步

+++ src/BiliLite.UWP/Api/Home/RecommendAPI.cs
@@ -16,7 +16,7 @@
             return api;
         }
 
-        public ApiModel Dislike(string _goto,string id,string mid,long reason_id,int rid,int tag_id)
+        public ApiModel Dislike(string _goto,string id,string mid,long reason_id,long rid,long tag_id)
         {
             ApiModel api = new ApiModel()
             {
@@ -27,7 +27,7 @@
             api.parameter += ApiHelper.GetSign(api.parameter, ApiHelper.AndroidKey);
             return api;
         }
-        public ApiModel Feedback(string _goto, string id, string mid, long feedback_id, int rid, int tag_id)
+        public ApiModel Feedback(string _goto, string id, string mid, long feedback_id, long rid, long tag_id)
         {
             ApiModel api = new ApiModel()
             {
